refactor: 상품 수정 데이터가 없을 경우 null로 데이터를 처리하도록 코드 리팩토링 #63

+++ be/src/main/java/team05a/secondhand/image/repository/ImageRepository.java
@@ -10,7 +10,7 @@ import team05a.secondhand.product.data.entity.Product;
 
 public interface ImageRepository extends JpaRepository<ProductImage, Long> {
 
-	Long countByProductId(Long productId);
+	int countByProductId(Long productId);
 
 	Optional<ProductImage> findFirstByProductId(Long productId);
